requirements.txt
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
apscheduler==3.10.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
awscli==1.33.33
babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.123
botocore==1.34.151
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cython==3.0.0
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docutils==0.16
email-validator==2.2.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gputil==1.4.0
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.2
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
langchain==0.2.11
langchain-core==0.2.26
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
langsmith==0.1.98
lazy-loader==0.4
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook-shim==0.2.4
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.15.1
openai==1.37.1
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.84
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pendulum==3.0.0
piexif==1.1.3
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.2
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.0.0
prompt-toolkit==3.0.47
psutil==6.0.0
pulp==2.9.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
pydeck==0.9.1
pydot==3.0.1
pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
pysocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
qtpy==2.4.1
redis==5.0.8
referencing==0.35.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.36
rpds-py==0.19.1
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.23.1
send2trash==1.8.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
skypilot==0.5.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlalchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
supervision==0.22.0
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
time-machine==2.14.2
tinycss2==1.3.0
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typer
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uvicorn==0.30.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
widgetsnbextension==4.0.11
wsproto==1.2.0
yarl==1.9.4
inference==0.15.2
